♻️ refactor: 프로필 페이지 경로 별칭 수정
